# contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Canchitas API",
    "version": "1.0.0",
    "description": "Contrato generado desde los esquemas zod (ADR 0007). No se edita a mano."
  },
  "components": {
    "securitySchemes": {
      "sesionWeb": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Secure-canchitas.session_token",
        "description": "Cookie HttpOnly, Secure y SameSite=Lax de la web (ADR 0009)"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token de Android: 30 días desde el último uso (RNF-012)"
      }
    },
    "schemas": {
      "SolicitudDeAltaInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "description": "Mail de la cuenta"
          },
          "contrasena": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "description": "Entre 8 y 128 caracteres"
          },
          "nombreUsuario": {
            "type": "string",
            "maxLength": 40,
            "description": "3 a 20 letras minúsculas, números, punto o guion bajo (RF-003). Se guarda en minúsculas."
          },
          "fechaNacimiento": {
            "type": "string",
            "format": "date",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))$",
            "description": "AAAA-MM-DD (RF-002)"
          },
          "aceptaPrivacidad": {
            "type": "boolean",
            "description": "Aceptación de la política de privacidad (RNF-018)"
          }
        },
        "required": [
          "email",
          "contrasena",
          "nombreUsuario",
          "fechaNacimiento",
          "aceptaPrivacidad"
        ]
      },
      "SolicitudDeInicioInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "description": "Mail de la cuenta"
          },
          "contrasena": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        },
        "required": [
          "email",
          "contrasena"
        ]
      },
      "SolicitudConEmailInput": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email",
            "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$",
            "description": "Mail de la cuenta"
          }
        },
        "required": [
          "email"
        ]
      },
      "SolicitudDeVerificacionInput": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096,
            "description": "Token del enlace del mail"
          }
        },
        "required": [
          "token"
        ]
      },
      "SolicitudDeRestablecimientoInput": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4096,
            "description": "Token del enlace del mail"
          },
          "contrasenaNueva": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "token",
          "contrasenaNueva"
        ]
      },
      "Problem": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "format": "uri"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "requisito": {
            "description": "ID del requisito del SRS, si aplica",
            "type": "string"
          },
          "errores": {
            "description": "Detalle por campo cuando la solicitud no es válida",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "campo": {
                  "type": "string"
                },
                "mensaje": {
                  "type": "string"
                }
              },
              "required": [
                "campo",
                "mensaje"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "type",
          "title",
          "status"
        ],
        "additionalProperties": false
      },
      "EstadoDeCuenta": {
        "type": "string",
        "enum": [
          "sin_verificar",
          "activa"
        ]
      },
      "Cuenta": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "email": {
            "type": "string"
          },
          "nombreUsuario": {
            "type": "string"
          },
          "estado": {
            "$ref": "#/components/schemas/EstadoDeCuenta"
          }
        },
        "required": [
          "id",
          "email",
          "nombreUsuario",
          "estado"
        ],
        "additionalProperties": false
      },
      "CuentaCreada": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "nombreUsuario": {
            "type": "string"
          },
          "estado": {
            "type": "string",
            "const": "sin_verificar"
          }
        },
        "required": [
          "email",
          "nombreUsuario",
          "estado"
        ],
        "additionalProperties": false
      },
      "SesionConToken": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Bearer para `Authorization`. Vence a los 30 días sin uso (RNF-012)."
          },
          "cuenta": {
            "$ref": "#/components/schemas/Cuenta"
          }
        },
        "required": [
          "token",
          "cuenta"
        ],
        "additionalProperties": false
      },
      "Salud": {
        "type": "object",
        "properties": {
          "estado": {
            "type": "string",
            "const": "ok"
          },
          "version": {
            "type": "string",
            "description": "Versión desplegada de la API"
          },
          "instante": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z|([+-](?:[01]\\d|2[0-3]):[0-5]\\d)))$",
            "description": "Hora del servidor"
          }
        },
        "required": [
          "estado",
          "version",
          "instante"
        ],
        "additionalProperties": false
      }
    }
  },
  "paths": {
    "/v1/salud": {
      "get": {
        "operationId": "consultarSalud",
        "summary": "Estado de la API",
        "tags": [
          "salud"
        ],
        "description": "Responde si la API está en línea y puede consultar la base. Hito M0 de docs/ROADMAP.md; no implementa un requisito del SRS.",
        "responses": {
          "200": {
            "description": "La API y la base responden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Salud"
                }
              }
            }
          },
          "503": {
            "description": "La base no responde",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cuentas": {
      "post": {
        "operationId": "registrarCuenta",
        "summary": "Crear una cuenta",
        "tags": [
          "cuentas"
        ],
        "description": "RF-001, RF-002, RF-003, RN-20 y RNF-018. Crea la cuenta en estado \"Sin verificar\" y manda el mail de verificación (RF-004). No inicia sesión.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudDeAltaInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cuenta creada sin verificar",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CuentaCreada"
                }
              }
            }
          },
          "409": {
            "description": "El mail (RF-001) o el nombre de usuario (RF-003) ya están en uso",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Menor de 18 años (RF-002), nombre de usuario o fecha inválidos, o sin aceptar la privacidad (RNF-018)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cuentas/sesion": {
      "post": {
        "operationId": "iniciarSesion",
        "summary": "Iniciar sesión en la web",
        "tags": [
          "cuentas"
        ],
        "description": "RF-005 y RNF-011. Deja la sesión en una cookie HttpOnly, Secure y SameSite=Lax (ADR 0009). Para Android se usa iniciarSesionConToken.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudDeInicioInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión iniciada; la cookie viaja en Set-Cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cuenta"
                }
              }
            }
          },
          "401": {
            "description": "El mail o la contraseña no son correctos, sin decir cuál (RF-005)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Inicio de sesión bloqueado por intentos fallidos (RNF-011). El encabezado Retry-After dice en cuántos segundos termina.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "cerrarSesion",
        "summary": "Cerrar la sesión del dispositivo actual",
        "tags": [
          "cuentas"
        ],
        "description": "RF-007. Borra la sesión (y el dispositivo registrado con ella, que deja de recibir push) y vence la cookie de la web.",
        "security": [
          {
            "sesionWeb": []
          },
          {
            "bearer": []
          }
        ],
        "responses": {
          "204": {
            "description": "Sesión cerrada"
          },
          "401": {
            "description": "No hay sesión",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cuentas/sesion/token": {
      "post": {
        "operationId": "iniciarSesionConToken",
        "summary": "Iniciar sesión en Android",
        "tags": [
          "cuentas"
        ],
        "description": "RF-005, RNF-011 y RNF-012. Devuelve un bearer que dura 30 días desde el último uso (ADR 0009). No deja cookie.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudDeInicioInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sesión iniciada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SesionConToken"
                }
              }
            }
          },
          "401": {
            "description": "El mail o la contraseña no son correctos, sin decir cuál (RF-005)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Inicio de sesión bloqueado por intentos fallidos (RNF-011). El encabezado Retry-After dice en cuántos segundos termina.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cuentas/yo": {
      "get": {
        "operationId": "consultarCuentaActual",
        "summary": "La cuenta de la sesión actual",
        "tags": [
          "cuentas"
        ],
        "description": "RF-004, RF-005 y RNF-012. Dice si hay sesión y si la cuenta está verificada. Usarla corre el vencimiento de la sesión (en la web, con una cookie nueva).",
        "security": [
          {
            "sesionWeb": []
          },
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Cuenta de la sesión",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cuenta"
                }
              }
            }
          },
          "401": {
            "description": "No hay sesión",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cuentas/verificacion": {
      "post": {
        "operationId": "verificarEmail",
        "summary": "Verificar el mail con el enlace",
        "tags": [
          "cuentas"
        ],
        "description": "RF-004 y RNF-014. Activa la cuenta. El enlace vence a las 24 horas y sirve una sola vez: si la cuenta ya está activa, se rechaza.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudDeVerificacionInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Cuenta activa"
          },
          "400": {
            "description": "El enlace venció o no es válido",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "410": {
            "description": "El enlace ya se usó (RNF-014)",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cuentas/verificacion/reenvio": {
      "post": {
        "operationId": "reenviarVerificacion",
        "summary": "Reenviar el mail de verificación",
        "tags": [
          "cuentas"
        ],
        "description": "RF-004. Responde igual exista o no la cuenta, para no revelar cuáles existen.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudConEmailInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Si la cuenta existe y no está verificada, se manda el mail"
          }
        }
      }
    },
    "/v1/cuentas/recuperacion": {
      "post": {
        "operationId": "pedirRecuperacion",
        "summary": "Pedir el mail para definir una contraseña nueva",
        "tags": [
          "cuentas"
        ],
        "description": "RF-006. Responde igual exista o no la cuenta. El enlace vence en 1 hora.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudConEmailInput"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Si la cuenta existe, se manda el mail"
          }
        }
      }
    },
    "/v1/cuentas/recuperacion/confirmacion": {
      "post": {
        "operationId": "restablecerContrasena",
        "summary": "Definir la contraseña nueva con el enlace",
        "tags": [
          "cuentas"
        ],
        "description": "RF-006 y RNF-014. El enlace sirve una sola vez. Cierra todas las sesiones abiertas de la cuenta.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudDeRestablecimientoInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Contraseña cambiada"
          },
          "400": {
            "description": "El enlace venció, no es válido o ya se usó",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        }
      }
    }
  }
}